requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
cloudpickle==3.1.2
    # via shap
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fonttools==4.62.1
    # via matplotlib
imbalanced-learn==0.14.1
    # via -r requirements.in
joblib==1.5.3
    # via
    #   imbalanced-learn
    #   scikit-learn
kiwisolver==1.5.0
    # via matplotlib
llvmlite==0.47.0
    # via
    #   numba
    #   shap
matplotlib==3.10.8
    # via
    #   -r requirements.in
    #   seaborn
numba==0.65.0
    # via shap
numpy==2.4.4
    # via
    #   -r requirements.in
    #   contourpy
    #   imbalanced-learn
    #   matplotlib
    #   numba
    #   pandas
    #   scikit-learn
    #   scipy
    #   seaborn
    #   shap
    #   xgboost
packaging==26.0
    # via
    #   matplotlib
    #   shap
pandas==3.0.2
    # via
    #   -r requirements.in
    #   seaborn
    #   shap
pillow==12.2.0
    # via matplotlib
pyarrow==23.0.1
    # via -r requirements.in
pyparsing==3.3.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
scikit-learn==1.8.0
    # via
    #   -r requirements.in
    #   imbalanced-learn
    #   shap
    #   sklearn-compat
scipy==1.17.1
    # via
    #   imbalanced-learn
    #   scikit-learn
    #   shap
    #   xgboost
seaborn==0.13.2
    # via -r requirements.in
shap==0.51.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sklearn-compat==0.1.5
    # via imbalanced-learn
slicer==0.0.8
    # via shap
threadpoolctl==3.6.0
    # via
    #   imbalanced-learn
    #   scikit-learn
tqdm==4.67.3
    # via shap
typing-extensions==4.15.0
    # via shap
xgboost==3.2.0
    # via -r requirements.in
